pyproject.toml
[tool.poetry]
name = "mixintune"
version = "0.1.0"
description = ""
authors = ["kalitka <zavorotnii.dr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
async-spotify = "^0.4.4"
black = "^23.10.1"
uvicorn = "^0.24.0.post1"
aioredis = "^2.0.1"
pydantic-settings = "^2.0.3"
fastapi-sso = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
